+++ Gw2Sharp/WebApi/V2/Clients/BaseEndpointBulkAllClient.cs
@@ -81,7 +81,7 @@ namespace Gw2Sharp.WebApi.V2.Clients
             this.RequestPageAsync<TObject, TId>(page, cancellationToken);
 
         /// <inheritdoc />
-        public virtual Task<IApiV2ObjectList<TObject>> PageAsync(int page, CancellationToken cancellationToken, int pageSize) =>
+        public virtual Task<IApiV2ObjectList<TObject>> PageAsync(int page, int pageSize, CancellationToken cancellationToken) =>
             this.RequestPageAsync<TObject, TId>(page, cancellationToken, pageSize);
     }
 }